requirements.txt
fastapi
uvicorn
httpx
authlib 
python-jose
